refactor(tasks): move stats logic to service and remove repo from controller
- Removes direct injection and usage of TaskRepository from TasksController.
- Creates TasksService.getStats method and moves calculation logic into it.
- Updates TasksController.getStats to inject request, retrieve user payload,
  and call TasksService.getStats, passing user context.
- Adds basic role check in TasksService.getStats to filter tasks by user ID
  for non-admins (stats calculation still requires optimization).
